# doc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VCS Core API",
    "version": "1.0.0",
    "description": "Version Control System API \u2014 Organization and Document management."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    },
    {
      "url": "https://api.vcs-system.com/v1",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "status": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PageMeta": {
        "type": "object",
        "required": [
          "page",
          "size",
          "totalElements",
          "totalPages"
        ],
        "properties": {
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          },
          "totalElements": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      },
      "UserProfile": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "photo_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          }
        }
      },
      "UpdateUserProfileRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "photo_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          }
        }
      },
...
      "Version": {
        "type": "object",
        "required": [
          "id",
          "doc_id",
          "version_number",
          "status",
          "is_draft",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "doc_id": {
            "type": "string",
            "format": "uuid"
          },
          "version_number": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "is_draft": {
            "type": "boolean"
          },
          "checksum": {
            "type": "string"
          },
          "storage_type": {
            "type": "string",
            "enum": [
              "SNAPSHOT",
              "DIFF"
            ]
          },
          "diff_preview": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Comment": {
        "type": "object",
        "required": [
          "id",
          "version_id",
          "author_id",
          "content",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "version_id": {
            "type": "string",
            "format": "uuid"
          },
          "author_id": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "author_name": {
            "type": "string",
            "description": "Display name of the comment author."
          }
        }
      },
      "UserSearchResult": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        }
      },
      "S3UploadResponse": {
        "type": "object",
        "required": [
          "upload_url"
        ],
        "properties": {
          "upload_url": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
      "DiffResponse": {
        "type": "object",
        "required": [
          "from_version_id",
          "to_version_id",
          "diff"
        ],
        "properties": {
          "from_version_id": {
            "type": "string",
            "format": "uuid"
          },
          "to_version_id": {
            "type": "string",
            "format": "uuid"
          },
          "diff": {
            "type": "string"
          }
        }
      },
      "PagedVersions": {
        "type": "object",
        "required": [
          "content",
          "meta"
        ],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Version"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        }
      },
...
      "PendingReviewItem": {
        "type": "object",
        "required": [
          "version_id",
          "doc_id",
          "doc_name",
          "author_id",
          "version_number",
          "created_at"
        ],
        "properties": {
          "version_id": {
            "type": "string",
            "format": "uuid"
          },
          "doc_id": {
            "type": "string",
            "format": "uuid"
          },
          "doc_name": {
            "type": "string"
          },
          "author_id": {
            "type": "string",
            "format": "uuid"
          },
          "version_number": {
            "type": "integer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/users/search": {
      "get": {
        "operationId": "searchUsers",
        "summary": "Search users by name or email",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search query (name or email)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSearchResult"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/user/profile": {
      "get": {
        "operationId": "getUserProfile",
        "summary": "Get current user profile",
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "patch": {
        "operationId": "updateUserProfile",
        "summary": "Update current user profile",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/documents/{doc_id}/versions": {
      "parameters": [
        {
          "name": "doc_id",
          "in": "path",
          "required": true,
          "description": "Document UUID",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "listVersions",
        "summary": "List version history (paginated)",
        "tags": [
          "Versions"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of versions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedVersions"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "operationId": "createVersion",
        "summary": "Create new version and get S3 upload URL",
        "tags": [
          "Versions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "checksum",
                  "is_draft",
                  "is_diff"
                ],
                "properties": {
                  "checksum": {
                    "type": "string"
                  },
                  "is_draft": {
                    "type": "boolean"
                  },
                  "is_diff": {
                    "type": "boolean"
                  },
                  "diff_preview": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created \u2014 S3 upload URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/S3UploadResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/documents/{doc_id}/versions/diff": {
      "parameters": [
        {
          "name": "doc_id",
          "in": "path",
          "required": true,
          "description": "Document UUID",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getVersionDiff",
        "summary": "Get diff between two versions",
        "tags": [
          "Versions"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "description": "Base version UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "description": "Target version UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Diff between two versions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/documents/{doc_id}/versions/{version_id}": {
      "parameters": [
        {
          "name": "doc_id",
          "in": "path",
          "required": true,
          "description": "Document UUID",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "version_id",
          "in": "path",
          "required": true,
          "description": "Version UUID",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getVersion",
        "summary": "Get version details",
        "tags": [
          "Versions"
        ],
        "responses": {
          "200": {
            "description": "Version details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Version"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/documents/{doc_id}/versions/{version_id}/download": {
      "parameters": [
        {
          "name": "doc_id",
          "in": "path",
          "required": true,
          "description": "Document UUID",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "version_id",
          "in": "path",
          "required": true,
          "description": "Version UUID",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getVersionDownloadUrl",
        "summary": "Get S3 pre-signed download URL for version file",
        "tags": [
          "Versions"
        ],
        "responses": {
          "200": {
            "description": "S3 download URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "download_url"
                  ],
                  "properties": {
                    "download_url": {
                      "type": "string",
                      "format": "uri"
                    }
                  }
                }
              }
            }
          },
          "202" : {
            "description": "Wait for notification with download URL",
            "content" : {}
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/documents/{doc_id}/versions/{version_id}/rollback": {
      "parameters": [
        {
          "name": "doc_id",
          "in": "path",
          "required": true,
          "description": "Document UUID",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "version_id",
          "in": "path",
          "required": true,
          "description": "Version UUID to roll back to",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "rollbackVersion",
        "summary": "Rollback \u2014 creates a new version that is a copy of the target",
        "tags": [
          "Versions"
        ],
        "responses": {
          "201": {
            "description": "New version created as a copy of the target",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Version"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/documents/{doc_id}/versions/{version_id}/approve": {
      "parameters": [
        {
          "name": "doc_id",
          "in": "path",
          "required": true,
          "description": "Document UUID",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "version_id",
          "in": "path",
          "required": true,
          "description": "Version UUID",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "approveVersion",
        "summary": "Approve version (reviewer only)",
        "tags": [
          "Versions"
        ],
        "responses": {
          "200": {
            "description": "Version approved"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/documents/{doc_id}/versions/{version_id}/reject": {
      "parameters": [
        {
          "name": "doc_id",
          "in": "path",
          "required": true,
          "description": "Document UUID",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "version_id",
          "in": "path",
          "required": true,
          "description": "Version UUID",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "rejectVersion",
        "summary": "Reject version (reviewer only)",
        "tags": [
          "Versions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Version rejected"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/documents/{doc_id}/versions/{version_id}/comments": {
      "parameters": [
        {
          "name": "doc_id",
          "in": "path",
          "required": true,
          "description": "Document UUID",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "version_id",
          "in": "path",
          "required": true,
          "description": "Version UUID",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "listComments",
        "summary": "List comments on a version",
        "tags": [
          "Versions"
        ],
        "responses": {
          "200": {
            "description": "List of comments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "operationId": "addComment",
        "summary": "Add comment on a version (reviewer)",
        "tags": [
          "Versions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/versions/pending-review": {
      "get": {
        "operationId": "listPendingReviewVersions",
        "summary": "List versions pending review assigned to current user",
        "tags": [
          "Versions"
        ],
        "responses": {
          "200": {
            "description": "Versions awaiting review by the authenticated reviewer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PendingReviewItem"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  }
}